Add files via upload
The blocked force data for a vacuum-powered, 3D-printed fingertip haptic device. The data is collected using a Nano-17 force sensor. The data consists of 6-DoF force and torque measurements. 

V corresponds to the voltage input to a vacuum regulator and the number next to it is the voltage level. Higher the voltage value higher the force output of the device. 